refactor(js): move force-external scripts from modules/ to scripts/
